refactor(server): move source to src and update build
