+++ server/package.json
@@ -3,8 +3,8 @@
   "version": "1.0.0",
   "type": "module",
   "scripts": {
-    "dev": "NODE_ENV=development tsx index.ts",
-    "build": "tsc --outDir dist && tscpaths -p tsconfig.json -s . -o ./dist",
+    "dev": "NODE_ENV=development tsx src/index.ts",
+    "build": "tsc --outDir dist && tscpaths -p tsconfig.json -s ./src -o ./dist",
     "start": "NODE_ENV=production node dist/index.js",
     "check": "tsc --noEmit",
     "main": "dist/index.js"
