+++ server/src/index.ts
@@ -2,7 +2,7 @@ import "dotenv/config";
 import express, { type Request, Response, NextFunction } from "express";
 import { registerRoutes } from "./routes.js";
 // import router from "./router.js"; // Removed unused import
-import { setupVite, serveStatic, log } from "./vite-runtime";
+import { setupVite, serveStatic, log } from "./vite-runtime.js";
 // Initialize application with MongoDB
 console.log('Initializing MarrakechDunes with MongoDB Atlas...');
