+++ server/src/security-middleware.ts
